+++ odoo/local-src/sf_account/models/account_invoice.py
@@ -2,6 +2,11 @@
 # Part of sensefly.
 
 from odoo import fields, models, api
+from odoo.exceptions import ValidationError
+import json
+import logging
+
+_logger = logging.getLogger(__name__)
 
 
 class AccountInvoice(models.Model):
@@ -53,6 +58,21 @@ class AccountInvoice(models.Model):
         self.sent = True
         return self.env['report'].get_action(
             self, 'sf_account.sf_report_invoice')
+
+    @api.one
+    def get_payment_info(self):
+        # Just load JSON payment info
+        json_data = self.payments_widget
+        try:
+            payment_info = json.loads(json_data)
+        except Exception, e:
+            log_msg = "Error loading payment info. %s." % (e.message,)
+            _logger.exception(log_msg)
+            raise ValidationError('Error loading payment info. %s' % e.message)
+
+        if payment_info:
+            return payment_info.get('content', False)
+        return False
 
     order_ids = fields.Many2many(
         'sale.order', string='Orders', compute='_compute_sale_orders')