+++ odoo/local-src/sf_account_payment_order/models/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 # Part of sensefly.
 from . import account_move_line
+from . import account_payment_order