+++ odoo/local-src/sf_partner/models/partner.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 # Part of SenseFly.
 from odoo import fields, models
+import uuid
 
 
 class Partner(models.Model):
@@ -8,3 +9,10 @@ class Partner(models.Model):
 
     is_intragroup = fields.Boolean(string='Is Intragroup',
                                    help='This partner belongs to Parrot Group')
+    uuid = fields.Char(
+        'UUID', index=True, default=lambda self: '%s' % uuid.uuid4(),
+        required=True)
+
+    _sql_constraints = [
+        ('unique_uuid', 'UNIQUE(uuid)', 'UUID must be unique!'),
+    ]