+++ odoo/local-src/sf_stock/models/stock_production_lot.py
@@ -3,12 +3,16 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from odoo import models, fields
+import uuid
 
 
 class ProductionLot(models.Model):
 
     _inherit = 'stock.production.lot'
 
+    uuid = fields.Char(
+        'UUID', index=True, default=lambda self: '%s' % uuid.uuid4(),
+        required=True)
     warranty_end_date = fields.Date('Warranty end date')
     first_outgoing_stock_move_id = fields.Many2one(
         'stock.move', string='First outgoing stock move')
@@ -17,3 +21,7 @@ class ProductionLot(models.Model):
                                         string='Product tracking')
 
     notes = fields.Text(string='Comment')
+
+    _sql_constraints = [
+        ('unique_uuid', 'UNIQUE(uuid)', 'UUID must be unique!'),
+    ]