+++ HISTORY.rst
@@ -26,6 +26,7 @@ latest (unreleased)
 **Bugfixes**
 
 * BSSFL-528: Account move form in edit mode do not present well team field
+* BSSFL-529: Down payment invoice number and related account move name in sync
 
 **Build**