+++ HISTORY.rst
@@ -22,6 +22,8 @@ latest (unreleased)
 **Features and Improvements**
 
 * BSSFL-375: Payment statement report
+* BIZ-1503: Set move date equal to Payment Execution when
+  Payment Execution Date Type == Fixed Date
 
 **Bugfixes**